Move config to a separate package-wide module
