+++ chandra_aca/__init__.py
@@ -1,6 +1,7 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 import ska_helpers
 
+from .config import conf
 from .transform import *
 
 __version__ = ska_helpers.get_version(__package__)
