+++ chandra_aca/transform.py
@@ -7,34 +7,11 @@ The transform modules includes:
 - Science target coordinate to ACA frame conversions
 """
 
-# Configuration for chandra_aca transform
-from astropy import config
-from astropy.config import ConfigNamespace
-
-
-class ConfigItem(config.ConfigItem):
-    rootname = "chandra_aca"
-
-
-class Conf(ConfigNamespace):
-    """
-    Configuration parameters for chandra_aca.transform.
-    """
-
-    use_legacy_coeffs = ConfigItem(
-        False,
-        "Use legacy coefficients for ACA transforms instead of new ground/flight values.",
-    )
-
-
-conf = Conf()
-
-
 import numba
 import numpy as np
 from Quaternion import Quat
 
-from chandra_aca import dark_model
+from chandra_aca import conf, dark_model
 
 ###################################################################################
 # Legacy coefficients to avoid regression diffs in other packages in cases where
@@ -304,7 +281,7 @@ def pixels_to_yagzag(
     -------
     (yang, zang) each vector of the same length as row/col
     """
-    use_legacy = conf.use_legacy_coeffs
+    use_legacy = conf.transform_use_legacy_coeffs
 
     row = np.asarray(row, dtype=np.float64)
     col = np.asarray(col, dtype=np.float64)
@@ -372,7 +349,7 @@ def yagzag_to_pixels(
     """
     yang = np.asarray(yang, dtype=np.float64)
     zang = np.asarray(zang, dtype=np.float64)
-    use_legacy = conf.use_legacy_coeffs
+    use_legacy = conf.transform_use_legacy_coeffs
 
     if use_legacy:
         row, col = _poly_convert(yang, zang, ACA2PIX_coeff, t_aca=t_aca)
